Introduce Clump class for handling Follow requests and improve validation logic
