+++ Server/KristofferStrube.ActivityPubBotDotNet.Server/Users/UsersApi.cs
@@ -5,70 +5,6 @@ using static System.Text.Json.JsonSerializer;
 
 namespace KristofferStrube.ActivityPubBotDotNet.Server;
 
-public class Clump
-{
-    private readonly IEnumerable<IObjectOrLink> _actor;
-    private readonly IEnumerable<IObjectOrLink> _obj;
-
-    private Clump(Follow follow,
-        string userUrl,
-        UserInfo user,
-        IEnumerable<IObjectOrLink> actor,
-        IEnumerable<IObjectOrLink> obj)
-    {
-        _actor = actor;
-        _obj = obj;
-        Follow = follow;
-        UserUrl = userUrl;
-        User = user;
-    }
-
-    public static Clump Create(Follow follow, string userUrl, UserInfo user)
-    {
-        IReadOnlyCollection<IObjectOrLink> actor = (follow.Actor ?? Enumerable.Empty<IObjectOrLink>()).ToArray();
-        IReadOnlyCollection<IObjectOrLink> obj = (follow.Object ?? Enumerable.Empty<IObjectOrLink>()).ToArray();
-
-        var errors = Validate(actor, userUrl, obj);
-        if (errors != null)
-            throw new ArgumentException(errors);
-
-        return new Clump(follow, userUrl, user, actor, obj);
-    }
-
-    public Follow Follow { get; }
-    public string UserUrl { get; }
-    public UserInfo User { get; }
-
-    private static string? Validate(IEnumerable<IObjectOrLink> actor, string userUrl, IEnumerable<IObjectOrLink> obj)
-    {
-        string? ret = null;
-        if (ActivityPubService.GetPersonId(obj.First()) is not { } objectPersonId)
-            ret = "The Object was not a Link or did not have a id.";
-        else if (objectPersonId != userUrl)
-            ret = "The Object Id did not match the address of this inbox.";
-        else if (ActivityPubService.GetPersonId(actor.First()) is null)
-            ret = "The Actor was not a Link or did not have a id.";
-        return ret;
-    }
-
-    public string GetPersonId() => ActivityPubService.GetPersonId(_obj.First())!;
-
-    public string FollowedId() => User.Id;
-
-    public IObjectOrLink Actor() => _actor.First();
-
-    public Accept AcceptPayload(IConfiguration configuration)
-    {
-        var accept = new Accept
-        {
-            Actor = new List<Link> { new() { Href = new(UserUrl) } },
-            Id = $"{configuration["HostUrls:Server"]}/Activity/{Guid.NewGuid()}",
-            Object = new List<IObject>() { Follow }
-        };
-        return accept;
-    }
-}
-
 public static class UsersApi
 {
     public static RouteGroupBuilder MapUsers(this IEndpointRouteBuilder routes)
